refactor: 상세 페이지 컴포넌트 분리
- poll-category-buttons, poll-option-button 컴포넌트로 분리
